+++ CHANGELOG.md
@@ -49,6 +49,11 @@ Changes to `GpuFuture`:
 Changes to pipelines:
 - The `descriptor_binding_requirements` method is moved to the `Pipeline` trait, and returns a reference to the hashmap directly.
 
+Changes to `impl_vertex`, `VertexMember` and `VertexMemberInfo`:
+- `VertexMember::format` now directly returns a `Format`
+- `VertexMemberInfo` now holds format and `VertexMemberTy` was removed.
+- `impl_vertex` was updated to correctly use the new `VertexMember::format` signature.
+
 ### Additions
 - Added `CpuBufferAllocatorCreateInfo`.
 - Allow waiting on `SwapchainAcquireFuture`.
@@ -60,6 +65,9 @@ Changes to pipelines:
 - `VertexMember` is now implemented for cgmath `Vector`s and `Point`s.
 - `type_for_format_cgmath` and `type_for_format_nalgebra` macros, next to the existing `type_for_format` macro.
 - Vulkano-shaders: `shader_cgmath` and `shader_nalgebra` macros, next to the existing `shader` macro.
+- Added a derive macro for the `Vertex` trait. The `impl_vertex` macro and `VertexMember` trait are deprecated.
+- `BufferDefinition` matching logic was updated to work with the new `VertexMemberInfo` and now matches based on scalar type, number of components and number of elements allowing the use of formats such as `*_UNORM`.
+- `GraphicsPipelineBuilder` validation was extended to make sure locations are not bound multiple times when a single attribute spans multiple locations (e.g. using double precision float formats).
 
 ### Bugs fixed
 - [#2094](https://github.com/vulkano-rs/vulkano/issues/2094): Fixed debug assertion when the first command in a command buffer that uses an image expects it to be in the `Undefined` layout.